+++ CHANGELOG.rst
@@ -4,8 +4,9 @@ xCORE-200 DSP library change log
 6.3.0
 -----
 
-  * ADDED: float4 FFT functions for float4 values
+  * ADDED: FFT functions for float4 values
   * ADDED: float FFT functions using lib_xs3_math, limited to 512 real points
+  * ADDED: float dot product functions for XS3
 
 6.2.1
 -----